+++ src/client/assets/favicon.png
