+++ src/client/main/main.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Main Page</title>
     <!-- link to CSS file -->
-    <link rel="stylesheet" href="main.css">
+    <link rel="stylesheet" href="/main/main.css">
 </head>
 
 <body>
