+++ src/client/product/product.html
@@ -5,18 +5,21 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Edit Listing</title>
-  <link rel="stylesheet" href="common.css"> <!-- Link to CSS file for this specific page -->
-  <link rel="stylesheet" href="seller.css"> <!-- Link to CSS file for this specific page -->
+  <link rel="stylesheet" href="/common.css"> <!-- Link to CSS file for this specific page -->
+  <link rel="stylesheet" href="/seller/seller.css"> <!-- Link to CSS file for this specific page -->
 </head>
 
 <body>
   <div id="carousel">
     <div id="carousel-images">
+        
     </div>
     <div id="carousel-controls">
       <svg xmlns="http://www.w3.org/2000/svg" id="carousel-left-arrow" viewBox="0 0 320 512"><!--!Font Awesome Free 6.5.2 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.--><path d="M9.4 233.4c-12.5 12.5-12.5 32.8 0 45.3l192 192c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3L77.3 256 246.6 86.6c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0l-192 192z"/></svg>
       <svg xmlns="http://www.w3.org/2000/svg" id="carousel-right-arrow" viewBox="0 0 320 512"><!--!Font Awesome Free 6.5.2 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.--><path d="M310.6 233.4c12.5 12.5 12.5 32.8 0 45.3l-192 192c-12.5 12.5-32.8 12.5-45.3 0s-12.5-32.8 0-45.3L242.7 256 73.4 86.6c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0l192 192z"/></svg>
     </div>
+    <!-- Button for adding images -->
+    <button id="add-image-button">Add Image</button>
   </div>
   <div id="description">
     <div>
@@ -38,4 +41,6 @@
       Seller Email: <span id="seller-email-label"></span>
     </div>
   </div>
+  <script src="product.js" type="module"></script> <!-- Link to JavaScript file -->
 </body>
+</html>
