+++ src/client/profile/profile.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Login Page</title>
-  <link rel="stylesheet" href="common.css"> <!-- Link to CSS file for this specific page -->
+  <link rel="stylesheet" href="/common.css"> <!-- Link to CSS file for this specific page -->
 </head>
 <body>
     <div class="red-background">
