+++ src/client/seller/seller.js
@@ -1,4 +1,4 @@
-import { blobToURL, getListing } from "./api.js";
+import { blobToURL, getListing, putListing } from "../api.js";
 
 export async function onNavigate() {
     const searchParams = new URLSearchParams(window.location.search);
@@ -10,26 +10,42 @@ export async function onNavigate() {
     const carouselRightArrow = document.getElementById("carousel-right-arrow");
 
     const carouselImageList = [];
-    /**
-     * Adds a specific element to the carousel
-     * @param { HTMLElement } element 
-     */
-    function appendToCarousel(element) {
-        carouselImageContainer.appendChild(element);
-        carouselImageList.push(element);
-    }
+
+    // Append all current images
     for (const imageBlob of currentListing.carousel) {
         /** @type { HTMLImageElement } */
         const carouselImage = document.createElement("img");
         carouselImage.src = await blobToURL(imageBlob);
         carouselImage.classList.add("carousel-image");
-        appendToCarousel(carouselImage);
+        carouselImageContainer.appendChild(carouselImage);
+        carouselImageList.push(carouselImage);
     }
+
     // Append placeholder for adding image
     const newImagePlaceholderElement = document.createElement("div");
     newImagePlaceholderElement.id = "new-image-placeholder";
     newImagePlaceholderElement.innerHTML = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><!--!Font Awesome Free 6.5.2 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.--><path d="M256 512A256 256 0 1 0 256 0a256 256 0 1 0 0 512zM232 344V280H168c-13.3 0-24-10.7-24-24s10.7-24 24-24h64V168c0-13.3 10.7-24 24-24s24 10.7 24 24v64h64c13.3 0 24 10.7 24 24s-10.7 24-24 24H280v64c0 13.3-10.7 24-24 24s-24-10.7-24-24z"/></svg>`
-    appendToCarousel(newImagePlaceholderElement);
+    carouselImageContainer.appendChild(newImagePlaceholderElement);
+    carouselImageList.push(newImagePlaceholderElement);
+
+    // Add listener for file drops
+    newImagePlaceholderElement.addEventListener("dragenter", e => e.preventDefault());
+    newImagePlaceholderElement.addEventListener("dragover", e => e.preventDefault());
+    newImagePlaceholderElement.addEventListener("drop", async e => {
+        const newImageList = [...e.dataTransfer.files].filter(file => file.type.split("/")[0] === "image");
+        currentListing.carousel.push(...newImageList);
+        e.preventDefault();
+        for (const imageBlob of newImageList) {
+            /** @type { HTMLImageElement } */
+            const carouselImage = document.createElement("img");
+            carouselImage.src = await blobToURL(imageBlob);
+            carouselImage.classList.add("carousel-image");
+            carouselImageContainer.insertBefore(carouselImage, newImagePlaceholderElement);
+            carouselImageList.splice(carouselImageList.length - 1, 0, carouselImage);
+        }
+        await putListing(currentListing);
+        console.log(await getListing(currentListing._id))
+    });
 
     /**
      * Jumps to a specific image index
@@ -37,6 +53,7 @@ export async function onNavigate() {
      */
     function jumpToImage(index) {
         const imageWidthList = carouselImageList.map(image => image.getBoundingClientRect().width);
+        console.log(imageWidthList);
         const prevWidths = imageWidthList.slice(0, index);
         const pageWidth = carouselDiv.getBoundingClientRect().width;
 
@@ -47,16 +64,17 @@ export async function onNavigate() {
         carouselImageContainer.style.left = `${offset}px`;
     }
 
+    //Jump to new image block
     let currentImage = carouselImageList.length - 1;
-    carouselImageList.forEach(element => element.addEventListener("load", () => {
-        jumpToImage(currentImage)
-    }))
+    setTimeout(() => jumpToImage(currentImage), 100);
+
     carouselLeftArrow.addEventListener("click", () => {
         currentImage = Math.max(0, currentImage - 1);
         jumpToImage(currentImage)
-    })
+    });
+
     carouselRightArrow.addEventListener("click", () => {
         currentImage = Math.min(carouselImageList.length - 1, currentImage + 1);
         jumpToImage(currentImage)
-    })
+    });
 }
